#20 Rename submodule (api → core)
